requirements.txt
sounddevice
scipy
numpy
websockets
deep-translator
python-dotenv
